--- src/api/request/TaskUpdateRequest.java
@@ -1,33 +1,0 @@
-package api.request;
-
-import model.TaskStatus;
-
-public class TaskUpdateRequest {
-    private String taskDescription;
-    private String taskName;
-    private TaskStatus status;
-
-    public String getTaskDescription() {
-        return taskDescription;
-    }
-
-    public void setTaskDescription(String taskDescription) {
-        this.taskDescription = taskDescription;
-    }
-
-    public String getTaskName() {
-        return taskName;
-    }
-
-    public void setTaskName(String taskName) {
-        this.taskName = taskName;
-    }
-
-    public TaskStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(TaskStatus status) {
-        this.status = status;
-    }
-}
