--- src/api/router/BaseHttpHandler.java
@@ -1,29 +1,0 @@
-package api.router;
-
-import com.sun.net.httpserver.HttpExchange;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-public class BaseHttpHandler {
-    protected void sendText(HttpExchange exchange, String text) throws IOException {
-        byte[] response = text.getBytes(StandardCharsets.UTF_8);
-        exchange.getResponseHeaders().add("Content-Type", "application/json;charset=utf-8");
-        exchange.sendResponseHeaders(200, response.length);
-        exchange.getResponseBody().write(response);
-        exchange.close();
-    }
-
-    protected void sendNotFound(HttpExchange exchange) throws IOException {
-        String response = "404 Not Found";
-        exchange.sendResponseHeaders(404, response.length());
-        exchange.getResponseBody().write(response.getBytes(StandardCharsets.UTF_8));
-        exchange.close();
-    }
-
-    protected void sendError(HttpExchange exchange, String errorMessage) throws IOException {
-        exchange.sendResponseHeaders(400, errorMessage.length());
-        exchange.getResponseBody().write(errorMessage.getBytes(StandardCharsets.UTF_8));
-        exchange.close();
-    }
-}
